+++ kartothek/serialization/_generic.py
@@ -187,7 +187,7 @@ def filter_df_from_predicates(df, predicates, strict_date_types=False):
     -------
     pd.DataFrame
     """
-    indexer = np.broadcast_to(False, len(df))
+    indexer = np.zeros(len(df), dtype=bool)
     for conjunction in predicates:
         inner_indexer = np.ones(len(df), dtype=bool)
         for column, op, value in conjunction:
@@ -201,6 +201,74 @@ def filter_df_from_predicates(df, predicates, strict_date_types=False):
             )
         indexer = inner_indexer | indexer
     return df[indexer]
+
+
+def _ensure_type_stability(array_like, value, strict_date_types, require_ordered):
+    """
+    Ensure that the provided value and the provided array will have compatible
+    types, such that comparisons are unambiguous.
+
+    The type check is based on the numpy type system and accesses the arrays
+    `kind` attribute and asserts equality. The provided value will be
+    interpreted as a scalar in this case. For scalars which do not have a proper
+    python representation, we will relax the strictness as long as there is a
+    valid and unambiguous interpretation of a comparison operation. In
+    particular we consider the following combinations valid:
+
+        * unsigned integer (u) <> integer (i)
+        * zero-terminated bytes (S) <> Python Object (O)
+        * Unicode string (U) <> Python Object (O)
+
+    Parameters
+    ----------
+    strict_date_types: bool
+        If False, assume that datetime.date and datetime.datetime are
+        compatible types. In this case, the value is cast appropriately
+    require_ordered: bool
+        Indicate if the operator to be evaluated will require a notion of
+        ordering. In the case of pd.Categorical we will then assume a
+        lexicographical ordering and cast the pd.CategoricalDtype accordingly
+    """
+    # datetime is a subclass of date which is why we need this double check
+    date_type_to_cast = datetime.datetime if strict_date_types else datetime.date
+    if isinstance(value, date_type_to_cast):
+        value = pd.Timestamp(value).to_datetime64()
+    elif is_list_like(value) and len(value) and isinstance(value[0], date_type_to_cast):
+        value = [pd.Timestamp(val).to_datetime64() for val in value]
+
+    value_dtype = pd.Series(value).dtype
+
+    if require_ordered and pd.api.types.is_categorical(array_like):
+        array_value_type = array_like.cat.categories.dtype
+        if array_like.cat.categories.is_monotonic:
+            array_like = array_like.cat.as_ordered()
+        else:
+            array_like = array_like.cat.reorder_categories(
+                array_like.cat.categories.sort_values(), ordered=True
+            )
+    else:
+        array_value_type = array_like.dtype
+
+    # NULL types might not be preserved well, so try to cast floats (pandas default type) to the value type
+    # Determine the type using the `kind` interface since this is common for a numpy array, pandas series and pandas extension arrays
+    if array_like.dtype.kind == "f" and np.isnan(array_like).all():
+        if array_like.dtype.kind != value_dtype.kind:
+            array_like = array_like.astype(value_dtype)
+            array_value_type = array_like.dtype
+
+    type_comp = {value_dtype.kind, array_value_type.kind}
+    if (
+        len(type_comp) > 1
+        # UINT and INT are accepted
+        and not type_comp == {"u", "i"}
+        # various string kinds
+        and not type_comp == {"S", "O"}
+        and not type_comp == {"U", "O"}
+    ):
+        raise TypeError(
+            f"Unexpected type encountered. Expected {array_value_type.kind} but got {value_dtype.kind}."
+        )
+    return array_like, value
 
 
 def filter_array_like(
@@ -230,29 +298,18 @@ def filter_array_like(
         mask = np.ones(len(array_like), dtype=bool)
 
     if out is None:
-        out = np.empty(len(array_like), dtype=bool)
+        out = np.zeros(len(array_like), dtype=bool)
 
-    # datetime is a subclass of date which is why we need this double check
-    date_type_to_cast = datetime.datetime if strict_date_types else datetime.date
-    if isinstance(value, date_type_to_cast):
-        value = pd.Timestamp(value).to_datetime64()
-    elif is_list_like(value) and len(value) and isinstance(value[0], date_type_to_cast):
-        value = [pd.Timestamp(val).to_datetime64() for val in value]
+    # In the case of an empty list, don't bother with evaluating types, etc.
+    if is_list_like(value) and len(value) == 0:
+        false_arr = np.zeros(len(array_like), dtype=bool)
+        np.logical_and(false_arr, mask, out=out)
+        return out
 
-    if pd.api.types.is_categorical(array_like) and (">" in op or "<" in op):
-        if array_like.cat.categories.is_monotonic:
-            array_like = array_like.cat.as_ordered()
-        else:
-            array_like = array_like.cat.reorder_categories(
-                array_like.cat.categories.sort_values(), ordered=True
-            )
-
-    # NULL types might not be preserved well, so try to cast floats (pandas default type) to the value type
-    # Determine the type using the `kind` interface since this is common for a numpy array, pandas series and pandas extension arrays
-    if array_like.dtype.kind == "f" and np.isnan(array_like).all():
-        value_dtype = np.array([value]).dtype
-        if array_like.dtype.kind != value_dtype.kind:
-            array_like = array_like.astype(value_dtype)
+    require_ordered = "<" in op or ">" in op
+    array_like, value = _ensure_type_stability(
+        array_like, value, strict_date_types, require_ordered
+    )
 
     with np.errstate(invalid="ignore"):
         if op == "==":